+++ FluidNC/esp32/esp32s3/Platform.h
@@ -10,7 +10,7 @@
 
 // The number that we support, regardless of how many the chip has
 #define MAX_N_DACS 0
-#define MAX_N_RMT 0
+#define MAX_N_RMT 4
 #define MAX_N_I2SO 1
 #define MAX_N_SPI 1
 #define MAX_N_SDCARD 1
